version 1/fake_banking_apk_detector/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0

# APK analysis
python-magic>=0.4.24
cryptography>=3.4.8

# Web framework
Flask>=2.0.0
Werkzeug>=2.0.0

# Configuration
PyYAML>=6.0

# Logging and utilities
colorlog>=6.6.0
tqdm>=4.62.0

# Optional: Advanced ML models
xgboost>=1.5.0
lightgbm>=3.3.0

# Optional: Deep learning (if implementing neural networks)
# tensorflow>=2.8.0
# torch>=1.10.0

# Development dependencies
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.0.0
flake8>=4.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
